+++ movieInsight/src/main/java/semi/project/movieInsight/cinema/dao/CinemaDetailDAO.java
@@ -149,10 +149,19 @@ public class CinemaDetailDAO {
 	 */
 	public float sumCinemaGrade(String cinemaName) {
 		
+//		Cinema sumcinema = sqlSession.selectOne("cinemaMapper.sumCinemaGrade",cinemaName);
+		
+		Float sumCinemaGrade = sqlSession.selectOne("cinemaMapper.sumCinemaGrade", cinemaName);
+
+		float sum = (sumCinemaGrade != null) ? sumCinemaGrade : 0.0f;
+
+	    System.out.println(sum);
+	
+		/*
 		float sum = sqlSession.selectOne("cinemaMapper.sumCinemaGrade",cinemaName);
-		
+	
 		System.out.println(sum);
-		
+		*/
 		return sum;
 	}
 
@@ -177,6 +186,15 @@ public class CinemaDetailDAO {
 		
 		
 		return cinemaList;
+	}
+
+
+	/**  영화관 댓글 수정
+	 * @param cinema
+	 * @return
+	 */
+	public int updateCommentCinema(Cinema cinema) {
+		return sqlSession.update("mypageMapper.updateCinemaComment", cinema);
 	}
