+++ movieInsight/src/main/webapp/resources/css/movie/movie-detail-page.css
@@ -1,10 +1,3 @@
-/*
-* {
-    box-sizing: border-box;
-}
-*/
-
-  
 
 body{
     margin: 0;
@@ -768,9 +761,6 @@ star{
     margin-left : 10px;
 }
 
-.comment-content{
-	white-space : pre-wrap;
-}
 
 
 
@@ -1741,7 +1731,7 @@ main>div{
 
 
 .comment-list-id{
-    flex-basis: 5%;
+    flex-basis: 7%;
 
    
 
@@ -1753,7 +1743,7 @@ main>div{
 }
 
 .comment-list-content{
-    flex-basis: 40%;
+    flex-basis: 38%;
 
     /* background-color: green; */
 
@@ -1761,8 +1751,14 @@ main>div{
     display: flex;
     justify-content: start;
     align-items: center;
-        font-weight : bold;
+    font-weight : bold;
 }
+
+/*
+ .comment-content{
+	white-space : pre-wrap;
+} 
+*/
 
 .comment-list-date{
     flex-basis: 30%;
@@ -1806,6 +1802,18 @@ margin-right: 30px;
     margin-right: 10px;
 }
 
+.saveBtn{
+    background-color: #383B3E;
+    color: rgb(228, 199, 137);
+    border: none;
+    border-radius: 10px;
+    padding: 10px 20px; 
+    cursor: pointer; 
+    font-size: 16px; 
+    margin: 0px;
+    margin-right: 30px;
+
+}
 .grade-container{
 	
     display : flex;
