+++ movieInsight/src/main/webapp/resources/css/mypage/mypage.css
@@ -760,6 +760,8 @@ aside > a{
 
 .comment-list-board {
 	flex-basis : 20%;
+	color : #3c60c2;
+	font-weight : bold;
 }
